Add simple split test

+++ yaramo/operations/split.py
@@ -4,7 +4,6 @@ from ..model import Edge, Topology
 
 
 class Split:
-
     @staticmethod
-    def split(topology: Topology, split_edges: Dict[Edge, int]) -> Tuple[Topology, Topology]:
+    def split(topology: Topology, split_edges: Dict[Edge, float]) -> Tuple[Topology, Topology]:
         return topology, topology
